moved solutions to files

+++ application/solution/repository/postgre.go
@@ -3,10 +3,14 @@ package repository
 import (
 	"context"
 	"errors"
+	"io/ioutil"
 	"liokoredu/application/models"
 	"liokoredu/application/solution"
+	"liokoredu/pkg/constants"
+	"liokoredu/pkg/generators"
 	"log"
 	"net/http"
+	"os"
 	"time"
 
 	"github.com/georgysavva/scany/pgxscan"
@@ -37,10 +41,30 @@ func (sd *SolutionDatabase) GetSolution(id uint64, taskId uint64, uid uint64) (m
 		return models.SolutionSQL{}, echo.NewHTTPError(http.StatusNotFound, "solution for task from user not found")
 	}
 
+	base, _ := os.Getwd()
+	sourceCode, _ := ioutil.ReadFile(base + constants.SolutionsDir + sln[0].SourceCode)
+	sln[0].SourceCode = string(sourceCode)
+
 	return sln[0], nil
 }
 
 func (sd *SolutionDatabase) DeleteSolution(id uint64, uid uint64) error {
+	var filename []string
+
+	err := pgxscan.Select(context.Background(), sd.pool, &filename,
+		`SELECT source_code from solutions WHERE id = $1 AND uid = $2`,
+		id, uid)
+
+	if err != nil {
+		log.Println("solution repo: DeleteSolution: error getting filename:", err)
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	if len(filename) == 0 {
+		log.Println("solution repo: DeleteSolution: error deleting solution: no file")
+		return echo.NewHTTPError(http.StatusNotFound, "solution with solutionId from this user not found")
+	}
+
 	resp, err := sd.pool.Exec(context.Background(),
 		`DELETE from solutions WHERE id = $1 AND uid = $2`,
 		id, uid)
@@ -55,6 +79,9 @@ func (sd *SolutionDatabase) DeleteSolution(id uint64, uid uint64) error {
 		return echo.NewHTTPError(http.StatusNotFound, "solution with solutionId from this user not found")
 	}
 
+	base, _ := os.Getwd()
+	_ = os.Remove(base + constants.SolutionsDir + filename[0])
+
 	return nil
 }
 
@@ -64,29 +91,31 @@ func (sd *SolutionDatabase) GetSolutions(taskId uint64, uid uint64) (models.Solu
 	err := pgxscan.Select(context.Background(), sd.pool, &sln,
 		`SELECT * FROM solutions WHERE task_id = $1 AND uid = $2`, taskId, uid)
 
-	if errors.As(err, &pgx.ErrNoRows) && len(sln) == 0 {
-		log.Println("solution repo: GetSolutions: error getting solution: no solutions")
-		return models.SolutionsSQL{}, nil
-	}
-
 	if err != nil {
 		log.Println("solution repo: GetSolutions: error getting solutions:", err)
 		return models.SolutionsSQL{}, echo.NewHTTPError(http.StatusInternalServerError, err.Error())
 	}
 
 	if len(sln) == 0 {
-		log.Println("solution repo: GetSolutions: error getting solution: no solutions")
+		log.Println("solution repo: GetSolutions: no solutions")
 		return models.SolutionsSQL{}, nil
+	}
+
+	base, _ := os.Getwd()
+	for _, elem := range sln {
+		sourceCode, _ := ioutil.ReadFile(base + constants.SolutionsDir + elem.SourceCode)
+		elem.SourceCode = string(sourceCode)
 	}
 
 	return sln, nil
 }
 
 // UpdateSolution implements solution.Repository
-func (sd *SolutionDatabase) UpdateSolution(id uint64, code int, tests int) error {
+func (sd *SolutionDatabase) UpdateSolution(id uint64, upd models.SolutionUpdate) error {
 	_, err := sd.pool.Exec(context.Background(),
-		`UPDATE solutions SET "check_result" = $1, "tests_passed" = $2 WHERE id = $3`,
-		code, tests, id)
+		`UPDATE solutions SET check_result = $1, tests_passed = $2, check_message = $3,
+		check_time = $4 WHERE id = $5`,
+		upd.Code, upd.Passed, upd.CheckMessage, upd.CheckTime, id)
 
 	if err != nil {
 		log.Println(err)
@@ -99,13 +128,31 @@ func (sd *SolutionDatabase) UpdateSolution(id uint64, code int, tests int) error
 func (sd *SolutionDatabase) InsertSolution(taskId uint64, uid uint64, code string, makefile string,
 	testsTotal int, receivedTime time.Time) (uint64, error) {
 	var id uint64
-	err := sd.pool.QueryRow(context.Background(),
+
+	filename := time.Now().Format("2006-01-02T15:04:05") +
+		generators.RandStringRunes(constants.PrivateLength)
+	data := []byte(code)
+
+	base, _ := os.Getwd()
+	file, err := os.Create(base + constants.SolutionsDir + filename)
+	if err != nil {
+		log.Println(err)
+	}
+	defer file.Close()
+
+	_, err = file.Write(data)
+	if err != nil {
+		log.Println("solution repo: InsertSolution: error creating file:", err)
+		return 0, echo.NewHTTPError(http.StatusBadRequest, err.Error())
+	}
+
+	err = sd.pool.QueryRow(context.Background(),
 		`INSERT INTO solutions (task_id, check_result, tests_passed, tests_total, 
 			received_date_time, source_code, uid, makefile) 
 		VALUES ($1, 1, 0, $2, $3, $4, $5, $6) RETURNING id`,
-		taskId, testsTotal, receivedTime, code, uid, makefile).Scan(&id)
+		taskId, testsTotal, receivedTime, filename, uid, makefile).Scan(&id)
 	if err != nil {
-		log.Println(err)
+		log.Println("solution repo: InsertSolution: error inserting solution", err)
 		return 0, echo.NewHTTPError(http.StatusBadRequest, err.Error())
 	}
